requirements.txt
requests
beautifulsoup4
fake-useragent
python-telegram-bot==20.0
Flask
